requirements.txt
# Empty Container DSS – Python dependencies

# Core data
pandas>=2.0
numpy>=1.24
pyarrow>=12.0        # Parquet I/O

# Validation
pydantic>=2.0
pandera>=0.17

# Forecasting
statsmodels>=0.14    # Holt-Winters, Theta
scikit-learn>=1.3    # RandomForest

# YAML config
pyyaml>=6.0

# Optimization
pulp>=2.7            # MILP solver (includes CBC)

# UI
streamlit>=1.32

# Testing
pytest>=7.4
